web_serv/requirements.txt
astroid==2.4.2
attrs==19.3.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
clickclick==1.2.2
connexion==2.7.0
Flask==1.1.2
Flask-Cors==3.0.8
google-api-core==1.22.0
google-api-python-client==1.11.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
google-cloud-core==1.3.0
google-cloud-firestore==1.8.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
httplib2==0.18.1
idna==2.10
importlib-metadata==1.7.0
inflection==0.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-spec-validator==0.2.9
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pylint==2.5.3
pyrsistent==0.16.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
swagger-ui-bundle==0.0.8
toml==0.10.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
